VERSION 0.8.0 medium fix: Do not die early
Before, ptybuffer immediately died when the command (child) terminated,
such that the connected sockets did not see the complete output.

This now has changed.  All buffered data is first delivered before
ptybuffer ends.  This is true for connected sockets when the child
dies unexpectedly or if the socket side is no more available.

If the child dies, everything up to the last output is delivered to
all connected sockets.  For now new connections are accepted
while delivery of data is waited for, but in future this will change,

This also means, that SIGHUP now can be delivered to the child:

If the socket side dies (with sockfile=- or if the listening socket
becomes unavailable) ptybuffer will wait until everything is
delivered to the child and then it closes the PTY, which in turn
sends a SIGHUP to the child.  It then will wait for the child
to exit.  Note that, if the child does not ignore SIGHUP, it
probably terminates with some default error.  This is standard.