VERSION 0.9.0 major bugfix and new option -w
This version terminates again if the PTY closes.  Version 0.8.0 was buggy in this respect (not in all use cases, but in the major ones).

Please note that ptybuffer before 0.8.0 died early without sending all data to all clients.
This now has changed. Ptybuffer does not terminate until the last client has got all the data.

Please also note that there is no lingering timeout yet for this.  So ptybuffer may stay indefinitely.
In future I might implement such a maximum lingering time, we will see.

There is a new option -w to control what happens, after the PTY closes:

Without option -w, the main socket is closed.  This means, no new clients can show up.
This is similar to before version 0.8.0, as ptybuffer terminated and thereby closed the main socket.
This is the intended behavior, such that option -c checks if the PTY is down and you are able to start a fresh command.

With option -w, the main socket is kept open until the last client has got all data.
This means, new clients can come in while the last one does not have all data.
This also means, that option -c does not check the PTY anymore, it only checks if ptybuffer is alive.