VERSION 0.10.0 Abstract Linux Sockets with @
The socket can now be abstract, which is a Linux feature.
Abstract sockets are specified starting with @

Also some other details were changed:

STDERR now gets the file log output of the child, too,
in case no logfile is chosen.  So logging to a logfile
or stderr now should contain the same thing
(except in one case: Option -l- sends child's stderr
to stderr instead to the tty.  This is for debugging.)