VERSION 0.12.0: IMPORTANT BUGFIX (option -g)
The TTY (actually stderr) was leaked to the forked child
in many circumstances.  This can even be a security problem,
in very special cases where some child assumes something else
and crosses some security border.

The FD is now correctly closed-on-exec.  If this fails,
there is a PROBLEM-note written to the log and ptybuffer stops
(previously this problem was silently ignored).

In the case that some weird environment is out there,
such problems can be ignored with option -g.
Then only a problem note is written to the log.