VERSION 0.13.0: option -r to open incoming TCP port
Option -r allows ptybuffer to open a TCP listening port.

Warning: TCP ports can be reachable from the world!

This allows a super simple chat like with

ptybuffer -r :6666 /bin/cat

Note:

This chat is not transparent, as the PTY eats several characters.
For example ^D is the EOF character and /bin/cat terminates on EOF.